pyproject.toml
[project]
name = "examen-dvc"
version = "1.0.0"
description = "Pipeline MLOps DVC pour l'examen de certification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dvc>=3.0",
    "dvc-s3>=3.0",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "numpy>=1.24"
]

[tool.uv]
dev-dependencies = []
